Fix issue with get_reply_to when ticket is 'fixed'
When a ticket is resolved (fixed) in snot. It is moved to the
/u/snot/[test]/spool/fixed dir instead of the /u/snot/[test]/spool/
active directory. For some bonghits reason, this breaks snot -m.

This adds logic to detect when -m is being called, and if the ticket
is archived, to unarchive it with snot -U, then get the message id.

It is safe to say that by the time that pysnot is asking for this
information, the ticket is about to be replied to anyways, so its
safe to do this.